+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Issues on the same line are now always displayed in the same order
+
 ## [0.2.0] - 2023-07-25
 
 ### Added